feat(tools): refactor + tool optionality
