+++ pkg/agent/loop.go
@@ -23,7 +23,6 @@ import (
 	"github.com/sipeed/picoclaw/pkg/logger"
 	"github.com/sipeed/picoclaw/pkg/providers"
 	"github.com/sipeed/picoclaw/pkg/routing"
-	"github.com/sipeed/picoclaw/pkg/skills"
 	"github.com/sipeed/picoclaw/pkg/state"
 	"github.com/sipeed/picoclaw/pkg/tools"
 	"github.com/sipeed/picoclaw/pkg/utils"
@@ -92,76 +91,25 @@ func registerSharedTools(
 			continue
 		}
 
-		// Web tools
-		if searchTool := tools.NewWebSearchTool(tools.WebSearchToolOptions{
-			BraveAPIKey:          cfg.Tools.Web.Brave.APIKey,
-			BraveMaxResults:      cfg.Tools.Web.Brave.MaxResults,
-			BraveEnabled:         cfg.Tools.Web.Brave.Enabled,
-			TavilyAPIKey:         cfg.Tools.Web.Tavily.APIKey,
-			TavilyBaseURL:        cfg.Tools.Web.Tavily.BaseURL,
-			TavilyMaxResults:     cfg.Tools.Web.Tavily.MaxResults,
-			TavilyEnabled:        cfg.Tools.Web.Tavily.Enabled,
-			DuckDuckGoMaxResults: cfg.Tools.Web.DuckDuckGo.MaxResults,
-			DuckDuckGoEnabled:    cfg.Tools.Web.DuckDuckGo.Enabled,
-			PerplexityAPIKey:     cfg.Tools.Web.Perplexity.APIKey,
-			PerplexityMaxResults: cfg.Tools.Web.Perplexity.MaxResults,
-			PerplexityEnabled:    cfg.Tools.Web.Perplexity.Enabled,
-		}); searchTool != nil {
-			agent.Tools.Register(searchTool)
-		}
-		if cfg.Tools.Core.EnableWebFetch {
-			agent.Tools.Register(tools.NewWebFetchTool(50000))
+		// specific context for this agent
+		agentCtx := tools.AgentContext{
+			AgentID:     agentID,
+			Workspace:   agent.Workspace,
+			Model:       agent.Model,
+			MaxTokens:   agent.MaxTokens,
+			Temperature: agent.Temperature,
 		}
 
-		// Hardware tools (I2C, SPI) - Linux only, returns error on other platforms
-		if cfg.Tools.Hardware.EnableI2C {
-			agent.Tools.Register(tools.NewI2CTool())
-		}
-		if cfg.Tools.Hardware.EnableSPI {
-			agent.Tools.Register(tools.NewSPITool())
+		// subagent security checker
+		currentAgentID := agentID
+		canSpawn := func(targetAgentID string) bool {
+			return registry.CanSpawnSubagent(currentAgentID, targetAgentID)
 		}
 
-		// Message tool
-		if cfg.Tools.Core.EnableMessage {
-			messageTool := tools.NewMessageTool()
-			messageTool.SetSendCallback(func(channel, chatID, content string) error {
-				msgBus.PublishOutbound(bus.OutboundMessage{
-					Channel: channel,
-					ChatID:  chatID,
-					Content: content,
-				})
-				return nil
-			})
-			agent.Tools.Register(messageTool)
-		}
+		// initialization
+		tools.SetupSharedTools(agent.Tools, cfg, msgBus, provider, agentCtx, canSpawn)
 
-		// Skill discovery and installation tools
-		if cfg.Tools.Skills.Enabled {
-			registryMgr := skills.NewRegistryManagerFromConfig(skills.RegistryConfig{
-				MaxConcurrentSearches: cfg.Tools.Skills.MaxConcurrentSearches,
-				ClawHub:               skills.ClawHubConfig(cfg.Tools.Skills.Registries.ClawHub),
-			})
-			searchCache := skills.NewSearchCache(
-				cfg.Tools.Skills.SearchCache.MaxSize,
-				time.Duration(cfg.Tools.Skills.SearchCache.TTLSeconds)*time.Second,
-			)
-			agent.Tools.Register(tools.NewFindSkillsTool(registryMgr, searchCache))
-			agent.Tools.Register(tools.NewInstallSkillTool(registryMgr, agent.Workspace))
-		}
-
-		// Spawn tool with allowlist checker
-		if cfg.Tools.Core.EnableSpawn {
-			subagentManager := tools.NewSubagentManager(provider, agent.Model, agent.Workspace, msgBus)
-			subagentManager.SetLLMOptions(agent.MaxTokens, agent.Temperature)
-			spawnTool := tools.NewSpawnTool(subagentManager)
-			currentAgentID := agentID
-			spawnTool.SetAllowlistChecker(func(targetAgentID string) bool {
-				return registry.CanSpawnSubagent(currentAgentID, targetAgentID)
-			})
-			agent.Tools.Register(spawnTool)
-		}
-
-		// Update context builder with the complete tools registry
+		// update context builder
 		agent.ContextBuilder.SetToolsRegistry(agent.Tools)
 	}
 }
